Prepares ssdp module parser for all HTTP messages.
Added Eupnp_HTTP_Response struct and renamed Eupnp_HTTP_Message to
Eupnp_HTTP_Request, as it is the appropriate name. Also renamed structs
attributes names for conforming with HTTP 1.1 RFC.

This prepares the parser for handling also HTTP responses. Next step is to
implement in fact the HTTP response parser - which differs basically on the
first line - and reuse actual header parser.